sieve: Install sieve scripts in a way that works with Dovecot
Dovecot requires that if the active sieve file is a symbolic link, it
must point to a non-symbolic link.  Since home-manager will create a
chain of links Dovecot fails to load the sieve file.

This change manually links in the sieve files during activation so
they point directly at the Nix store.  Dovecot seems to be fairly
happy about that.